requirements.txt
flask~=3.1.2
flask_session~=0.8.0

gunicorn~=23.0.0

requests~=2.32.0
beautifulsoup4~=4.14.0

python-dotenv~=1.2.0
anthropic~=0.75.0

pymysql~=1.1.0
DBUtils~=3.1.0

cryptography~=46.0.0
bcrypt~=5.0.0

PyYAML~=6.0.0